refactor: sanitized and refactored codebase
