+++ docs/home/installation.mdx
