+++ docs/home/tech-stack.mdx
