--- server/errors/index.ts
@@ -1,13 +1,0 @@
-import Bad_Request from './badRequest';
-import CustomError from './customError';
-import Forbidden from './forbidden';
-import Not_Found from './notFound';
-import Unauthorized from './unauthorized';
-
-export const ERR = {
-  CustomError,
-  Bad_Request,
-  Not_Found,
-  Unauthorized,
-  Forbidden,
-};
