+++ src/components/contributorpage/ContributorComp.tsx
@@ -1,7 +1,7 @@
-import type { TCont } from '&validation/contributor.validation';
+import { getFormattedDate } from '@/lib/Date';
+import type { TCont } from '@/server/models/validation/contributor.validation';
 import GithubIcon from '@icons/Github';
 import Glow from '@utilities/Glow';
-import { getFormattedDate } from 'lib/date';
 import Image from 'next/image';
 import type { ComponentProps, FC } from 'react';
