+++ src/context/HeaderContext.tsx
@@ -1,4 +1,4 @@
-import axios from 'config/axios';
+import axios from '@/config/axios';
 import {
   FC,
   ReactNode,
