+++ src/pages/_app.tsx
@@ -1,9 +1,9 @@
-import HeaderProvider from '@/context/headerContext';
-import ThemeProvider from '@/context/theme-context';
+import HeaderProvider from '@/context/HeaderContext';
+import ThemeProvider from '@/context/ThemeContext';
 import useProgressBar from '@/hooks/useProgressBar';
+import '@/styles/globals.css';
 import { Fira_Code, Inter } from '@next/font/google';
 import type { AppProps } from 'next/app';
-import '../styles/globals.css';
 
 const inter = Inter({
   display: 'block',
