+++ src/pages/api/v1/contributors/[contId].ts
@@ -2,8 +2,8 @@ import {
   deleteContributor,
   getSingleContributor,
   updateContributor,
-} from '&/controllers/contributors';
-import connectDB from '&/db/connectDB';
+} from '@/server/controllers/contributors';
+import connectDB from '@/server/db/connectDB';
 import { NextApiHandler } from 'next';
 
 // end-point: origin/api/v1/contributors
