+++ src/pages/contributors/index.tsx
@@ -2,13 +2,13 @@ import type { TCont } from '&validation/contributor.validation';
 import NewContributrs from '@/components/contributorpage/NewContributrs';
 import OldContributrs from '@/components/contributorpage/OldContributrs';
 import PopularContributrs from '@/components/contributorpage/PopularContributrs';
-import Container from '@layouts/Container';
-import Layout from '@layouts/Main';
 import {
   fetchNewContributors,
   fetchOldContributors,
   fetchPopularContributors,
-} from 'lib/contributors';
+} from '@/lib/Contributors';
+import Container from '@layouts/Container';
+import Layout from '@layouts/Main';
 import { GetStaticProps } from 'next';
 import type { FC } from 'react';
