+++ src/server/controllers/contributors.ts
@@ -4,7 +4,7 @@ import Contributor from '&/models/contributors';
 import { assertIsString } from '&/validator/assertionGuards';
 import type { TContributor } from '&validation/contributor.validation';
 import ContV from '&validation/contributor.validation';
-import { NextApiHandler } from 'next';
+import type { NextApiHandler } from 'next';
 
 function setUpdatableProperty<T, Prop extends keyof T>(
   value: object,
