+++ src/server/errors/customError.ts
