+++ src/server/errors/forbidden.ts
@@ -1,4 +1,4 @@
-import CustomError from './customError';
+import CustomError from '&/errors/customError';
 
 export default class Forbidden extends CustomError {
   constructor(message: string) {
