+++ src/server/errors/index.ts
@@ -1,0 +1,13 @@
+import Bad_Request from '&/errors/badRequest';
+import CustomError from '&/errors/customError';
+import Forbidden from '&/errors/forbidden';
+import Not_Found from '&/errors/notFound';
+import Unauthorized from '&/errors/unauthorized';
+
+export const ERR = {
+  CustomError,
+  Bad_Request,
+  Not_Found,
+  Unauthorized,
+  Forbidden,
+};
