+++ src/server/errors/notFound.ts
@@ -1,4 +1,4 @@
-import CustomError from './customError';
+import CustomError from '&/errors/customError';
 
 export default class Not_Found extends CustomError {
   constructor(message: string) {
