+++ src/server/errors/unauthorized.ts
@@ -1,4 +1,4 @@
-import CustomError from './customError';
+import CustomError from '&/errors/customError';
 
 export default class Unauthorized extends CustomError {
   constructor(message: string) {
