--- types/file.d.ts
@@ -1,6 +1,0 @@
-export declare type TFileContent = {
-  meta: {
-    [key: string]: unknown;
-  };
-  content: string;
-};
